Extract extensions and connectors
