+++ java/semantickernel-api/src/main/java/com/microsoft/semantickernel/memory/NullMemory.java
