+++ java/semantickernel-connectors-parent/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.microsoft.semantickernel</groupId>
+        <artifactId>semantickernel-parent</artifactId>
+        <version>0.2.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <packaging>pom</packaging>
+
+    <groupId>com.microsoft.semantickernel.connectors</groupId>
+    <artifactId>semantickernel-connectors-parent</artifactId>
+
+    <modules>
+        <module>semantickernel-connectors</module>
+    </modules>
+</project>
