+++ java/semantickernel-core/src/main/resources/META-INF/services/com.microsoft.semantickernel.orchestration.WritableContextVariables$Builder.java
@@ -1,0 +1,1 @@
+com.microsoft.semantickernel.orchestration.DefaultContextVariables$WritableBuilder
