+++ java/semantickernel-extensions-parent/sequentialplanner-extensions/src/main/java/com/microsoft/semantickernel/planner/SequentialPlannerRequestSettings.java
@@ -1,11 +1,10 @@
 // Copyright (c) Microsoft. All rights reserved.
 package com.microsoft.semantickernel.planner; // Copyright (c) Microsoft. All rights reserved.
 
+import javax.annotation.Nullable;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
-
-import javax.annotation.Nullable;
 
 /// <summary>
 /// Common configuration for planner instances.
