+++ samples/java/semantickernel-samples/pom.xml
@@ -30,8 +30,18 @@
     <dependencies>
         <dependency>
             <groupId>com.microsoft.semantickernel</groupId>
-            <artifactId>semantickernel-core</artifactId>
+            <artifactId>semantickernel-api</artifactId>
         </dependency>
+        <dependency>
+            <groupId>com.microsoft.semantickernel.connectors</groupId>
+            <artifactId>semantickernel-connectors</artifactId>
+            <version>0.2.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.microsoft.semantickernel</groupId>
+            <artifactId>openai-sdk</artifactId>
+        </dependency>
+
 
         <dependency>
             <groupId>org.apache.logging.log4j</groupId>
